Changing commands, variables, and operators.
Operator functions now take two RPNValues and return an RPNValue that is the
result of the operation. This reduced the high redundancy in the operator
functions.

Sqrt is now a command.

Variables module updated so it's freeing method is like that of commands and
operators.

Adding commands that take advantage of previously added "args" feature. However,
this commit's "unset" in broken.

	modified:   src/commands.c
	modified:   src/include/commands.h
	modified:   src/include/operators.h
	modified:   src/include/variables.h
	modified:   src/operators.c
	modified:   src/variables.c