Added fixes to make Visual C++ happy.
Visual C++ won't instantiate a template function when used as a
parameter to a function pointer. In other words, it doesn't like this:

    printAnyList(foo, printAnything);

So, I added some new functions, printValue() and printValueDetailed(),
and used them instead.

Also, it didn't like converting a 'void*' to a bool, which is what
Port::CanRun() does for the console port. So, I changed
'return std::cin;' to 'return !!std::cin;'.